# docs/api/mini-kasko-swagger.yaml
openapi: 3.0.3
info:
  title: MiniKasko Insurance API — UConnect
  version: 1.0.0
  description: |
    API для страхування MiniKasko (спрощене автокаско на суму покриття
    400 000 / 800 000 / 1 200 000 грн).

    ## Ключові відмінності від інших продуктів (Pets/Osago)

    - Немає `{mode}` (`test`/`prod`) у шляху — усі ендпоінти MiniKasko працюють
      без цього префіксу.
    - Двокроковий процес створення чернетки (`orders/init`) і заявлення
      страховику (`orders/{orderId}/declare`). Є також комбінований
      `POST /orders`, що виконує обидва кроки за один виклик.
    - Власний OTP-флоу (`send-otp` / `verify-otp` на замовленні) для підпису
      поліса — **не** спільний платіжний OTP.
    - Оплата, чек інвойсу та список платіжних систем — спільні ендпоінти,
      однакові з іншими продуктами (`/orders/get-invoice`,
      `/payments/check-invoice`, `/payment-systems/list`).
    - Завантаження документів — окремі **неавторизовані** GET-ендпоінти, що
      віддають бінарний файл напряму (не base64 у JSON, як у Pets).

    ## Загальний потік

    ```
    1.  POST /insurance/mini-kasko/calculate                 → розрахунок вартості, список пропозицій по СК
    2.  POST /insurance/mini-kasko/orders/init                → створення чернетки (statusId = 1)
        (декілька PUT /insurance/mini-kasko/orders/{orderId} для редагування чернетки)
    3.  POST /insurance/mini-kasko/orders/{orderId}/declare   → заявлення договору страховику (statusId = 5)
...
    ```

...
servers:
...
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Bearer токен авторизації. Передавати у заголовку:
        `Authorization: Bearer <token>`

        **Увага:** ендпоінти завантаження документів
        (`GET .../download`, `GET .../download-draft`) працюють **без токена** —
        доступ обмежується лише знанням `orderId`.
  responses:
    Unauthorized:
      description: Не авторизовано — Bearer токен відсутній або недійсний
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
          example:
            status: auth
            message: Authentication failed
...
  parameters:
    orderId:
      name: orderId
      in: path
      required: true
      description: ID замовлення. `id` з відповіді `/orders/init` або `/orders`
      schema:
        type: string
      example: "01JXXXXXXXXXXXXXXXXXXXXXXXXX"
  schemas:
    ApiErrorResponse:
      type: object
      description: Стандартна обгортка відповіді з помилкою
      properties:
        status:
          type: string
          description: |
            Категорія помилки:
            - `error` — загальна серверна помилка (HTTP 500) або помилка дії (HTTP 422)
            - `validation` — помилки валідації вхідних даних (HTTP 422). Деталі в полі `data`
            - `notFound` — запитаний ресурс не знайдено (HTTP 404)
            - `auth` — помилка автентифікації або авторизації (HTTP 401)
          enum: [error, validation, notFound, auth]
          example: error
        message:
          type: string
          description: Текстовий опис помилки
        data:
          type: object
          nullable: true
          description: |
            При `status=validation` — об'єкт, де ключі це назви полів із помилками,
            значення — масив текстів помилок валідації. При інших статусах — відсутнє.
    Car:
      type: object
      description: Дані транспортного засобу, що страхується
      required: [vin, year, is_taxi, reg_num, mark_name, model_name, category, registration_city_id]
      properties:
        vin:
          type: string
          description: VIN-код авто
          example: "WVWZZZ1KZAW123456"
        year:
          type: integer
          description: Рік випуску авто
          minimum: 1900
          maximum: 2100
          example: 2020
        is_taxi:
          type: integer
          description: Ознака використання як таксі. `1` — так, `0` — ні
          enum: [0, 1]
          example: 0
        reg_num:
          type: string
          description: Державний реєстраційний номер
          example: "AA1234BB"
        mark_name:
          type: string
          description: Марка авто
          example: "Skoda"
        model_name:
          type: string
          description: Модель авто
          example: "Octavia"
        category:
          type: string
          description: Категорія транспортного засобу
          example: "B"
        registration_city_id:
          type: integer
          description: ID міста реєстрації авто
          example: 14
        registration_city_name:
          type: string
          nullable: true
          description: Назва міста реєстрації для відображення. Необов'язкове
          example: "Харків"
    Customer:
      type: object
      description: Персональні дані страхувальника
      required:
        - code
        - surname
        - name
        - middle_name
        - phone
        - email
        - document_number
        - document_issued_by
        - document_issued_date
        - city_id
        - street
        - house
        - born_date
      properties:
        code:
          type: string
          description: |
            ІПН (10 цифр). Перевіряється контрольною сумою і має відповідати
            даті народження в `born_date` (дата народження зашифрована у
            перших 5 цифрах ІПН).
          example: "1234567890"
        surname:
          type: string
          example: "Петренко"
        name:
          type: string
          example: "Іван"
        middle_name:
          type: string
          example: "Іванович"
        phone:
          type: string
          example: "+380501112233"
        email:
          type: string
          format: email
          example: "user@example.com"
        document_number:
          type: string
          description: Номер паспорта
          example: "123456"
        document_issued_by:
          type: string
          description: Орган, що видав паспорт
          example: "Орган видачі"
        document_issued_date:
          type: string
          format: date
          example: "2015-06-01"
        city_id:
          type: integer
          description: ID міста страхувальника
          example: 14
        city_name:
          type: string
          nullable: true
          example: "Харків"
        street:
          type: string
          example: "Хрещатик"
        house:
          type: string
          example: "1"
        apartment:
          type: string
          nullable: true
          example: "5"
        born_date:
          type: string
          format: date
          description: Дата народження. Має збігатися з датою, зашифрованою в `code` (ІПН)
          example: "1990-01-01"
...
    OrderActionResult:
      type: object
      description: |
        Загальна форма успішної відповіді дій над замовленням
        (`init`, `orders`, `update`, `declare`, `replicate`). Точний набір
        полів залежить від конкретного ендпоінту — див. приклади відповідей
        для кожного шляху.
      properties:
        status:
          type: string
          enum: [success]
          example: success
        id:
          type: string
          description: ID замовлення. Використовувати у всіх наступних запитах
          example: "01JXXXXXXXXXXXXXXXXXXXXXXXXX"
        statusId:
          type: integer
          description: |
            Числовий статус замовлення:
            - `1` — draft (чернетка)
            - `2` — overdue (прострочене)
            - `5` — declared (заявлене страховику)
            - `6` — terminated (достроково розірване)
            - `10` — concluded (укладене/підтверджене)
            - `-10` — canceled
          enum: [1, 2, 5, 6, 10, -10]
          example: 5
        isEditable:
          type: boolean
          description: "`true` лише для `statusId = 1` (draft)"
          example: false
        isDeclared:
          type: boolean
          description: "`true` для будь-якого статусу окрім draft"
          example: true
        externalId:
          type: string
          nullable: true
          description: Зовнішній ID договору від страховика. Присутнє після `declare`
          example: "external-draft-1"
        policyNumber:
          type: string
          nullable: true
          description: Номер поліса від страховика. Присутнє після `declare`
          example: "PN-0001"
        isOfferValid:
          type: boolean
          description: "`false` якщо пропозиція, на якій базується чернетка, вже не дійсна"
          example: true
        isOfferExpired:
          type: boolean
          description: "`true` якщо пропозиція прострочена (дійсна лише в межах доби розрахунку)"
          example: false
        offerValidationMessage:
          type: string
          nullable: true
          description: Пояснення причини невалідності пропозиції, якщо `isOfferValid = false`
        clonedFromOrderId:
          type: string
          nullable: true
          description: ID замовлення-джерела. Присутнє лише у відповіді `/replicate`
          example: "01JXXXXXXXXXXXXXXXXXXXXXXXXX"
...
    CreateOrderRequest:
      type: object
      description: |
        Тіло для `POST /insurance/mini-kasko/orders/init` та
        `POST /insurance/mini-kasko/orders` (ідентична схема, різниця лише в
        тому, чи одразу виконується заявлення страховику).
      required: [offerId, car, dates, customer]
      properties:
        offerId:
          type: integer
          description: "`OfferSummary.offerId` з відповіді `/calculate`"
          example: 501
        car:
          $ref: '#/components/schemas/Car'
        dates:
          type: object
          required: [startDate]
          properties:
            startDate:
              type: string
              format: date
              description: Дата початку дії поліса
              example: "2026-06-20"
        customer:
          $ref: '#/components/schemas/Customer'
...
    SendOtpRequest:
      type: object
      required: [channel]
      properties:
        channel:
          type: string
          description: |
            Канал доставки OTP для підпису поліса:
            - `email` — лише email
            - `viber_sms` — email + Viber/SMS
          enum: [email, viber_sms]
          example: email
    VerifyOtpRequest:
      type: object
      required: [otp]
      properties:
        otp:
          type: string
          description: OTP, вбудований страховиком у відповідь заявлення договору
          example: "482913"
paths:
...
  /insurance/mini-kasko/orders:
    post:
      tags: [MiniKasko]
      summary: Створити замовлення (чернетка + заявлення за один виклик)
      description: |
        Комбінований ендпоінт: створює чернетку і одразу заявляє її
        страховику. Результат — заявлене страховику замовлення
        (`statusId = 5`), без проміжного кроку редагування чернетки.

        Використовувати, коли всі дані клієнта вже відомі і немає потреби
        показувати проміжний крок чернетки.
      operationId: createMiniKaskoOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '200':
          description: Замовлення створено та заявлено страховику
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderActionResult'
              example:
                status: success
                id: "01JXXXXXXXXXXXXXXXXXXXXXXXXX"
                statusId: 5
                isEditable: false
                isDeclared: true
                externalId: "external-draft-1"
                policyNumber: "PN-0001"
        '422':
          description: Помилка валідації або помилка дії
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /insurance/mini-kasko/orders/{orderId}/send-otp:
    post:
      tags: [MiniKasko]
      summary: Надіслати OTP для підпису поліса
      description: |
        Надсилає клієнту OTP код, необхідний для підпису поліса.

        Доступно лише для заявлених (declared) замовлень —
        `422 MiniKasko OTP can be sent only for declared orders.` інакше.

        `channel: email` — лист з прикріпленим PDF чернетки поліса.
        `channel: viber_sms` — email + Viber/SMS з текстом і посиланням на
        завантаження чернетки поліса.
      operationId: sendMiniKaskoOtp
      parameters:
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendOtpRequest'
      responses:
        '200':
          description: OTP надіслано
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "OTP sent."
        '422':
          description: |
            Помилка валідації (`channel`) або помилка дії — замовлення не
            заявлене, OTP недоступний, немає email/телефону клієнта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /insurance/mini-kasko/orders/{orderId}/verify-otp:
    post:
      tags: [MiniKasko]
      summary: Перевірити OTP для підпису поліса
      description: |
        Перевіряє введений клієнтом OTP код. При успіху OTP вважається
        підтвердженим — це джерело OTP для `/confirm`, якщо `otp` не
        передано напряму в тілі confirm-запиту.
      operationId: verifyMiniKaskoOtp
      parameters:
        - $ref: '#/components/parameters/orderId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpRequest'
      responses:
        '200':
          description: OTP вірний
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: "OTP verified."
        '422':
          description: "Невірний OTP (`MiniKasko OTP is invalid.`) або замовлення не заявлене"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...